requirements.txt
amadeus
fastapi
uvicorn
pydantic-settings
python-dotenv
sqlalchemy
alembic